+++ compat-assemblyadapter3/src/main/java/com/github/panpf/assemblyadapter3/compat/CompatExpandableChildItemFactory.kt
@@ -12,7 +12,7 @@ class CompatExpandableChildItemFactory(val itemFactory: CompatAssemblyItemFactor
         return itemFactory.match(data)
     }
 
-    override fun createItem(parent: ViewGroup): ExpandableChildItem<ExpandableGroup, Any> {
+    override fun createExpandableChildItem(parent: ViewGroup): ExpandableChildItem<ExpandableGroup, Any> {
         return CompatExpandableChildItem(itemFactory.dispatchCreateItem(parent))
     }